update: restructure component to WorkPage folder
