+++ src/components/WorkPage/CaseStudyHeroImage.tsx
