+++ src/components/WorkPage/index.tsx
@@ -2,7 +2,7 @@ import type { FC } from 'react';
 
 import { Layout } from '~/components';
 
-import { FeaturedCaseStudyHero } from '../FeaturedCaseStudyHero';
+import { FeaturedCaseStudyHero } from './FeaturedCaseStudyHero';
 
 export const WorkPage: FC = () => {
   return (
