+++ src/components/index.ts
@@ -3,7 +3,6 @@ export * from './Container';
 export * from './ContactPage';
 export * from './CulturePage';
 export * from './DynamicImage';
-export * from './FeaturedCaseStudyHero';
 export * from './GradientCircle';
 export * from './Header';
 export * from './Heading';
@@ -14,7 +13,6 @@ export * from './Layout';
 export * from './Link';
 export * from './MobileMenu';
 export * from './PageHero';
-export * from './CaseStudyHeroImage';
 export * from './QuotedImage';
 export * from './TeamCardList';
 export * from './TeamHero';
